pyproject.toml
[project]
name = "cdk-ollama"
version = "0.1.0"
description = ""
authors = [
    {name = "gon9a",email = "atsushi.tmail@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "constructs (>=10.4.2,<11.0.0)",
    "aws-cdk-lib (>=2.178.1,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
